requirements.txt
mysql-connector-python==8.3.0
schedule==1.2.1
openai==1.14.2
arxiv==2.1.0